Move reducer into react component
